Consolidate Metafile code
