+++ src/System.Drawing.Common/src/System/Drawing/Imaging/Metafile.Unix.cs
@@ -43,9 +43,7 @@ namespace System.Drawing.Imaging
 #if !NETCORE
     [Editor ("System.Drawing.Design.MetafileEditor, " + Consts.AssemblySystem_Drawing_Design, typeof (System.Drawing.Design.UITypeEditor))]
 #endif
-    [Serializable]
-    [System.Runtime.CompilerServices.TypeForwardedFrom("System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a")]
-    public sealed class Metafile : Image
+    public sealed partial class Metafile : Image
     {
 
         // constructors
@@ -74,113 +72,14 @@ namespace System.Drawing.Imaging
             Gdip.CheckStatus(status);
         }
 
-        public Metafile(string filename)
-        {
-            // Called in order to emulate exception behavior from netfx related to invalid file paths.
-            Path.GetFullPath(filename);
-
-            int status = Gdip.GdipCreateMetafileFromFile(filename, out nativeImage);
-            if (status == Gdip.GenericError)
-                throw new ExternalException("Couldn't load specified file.");
-            Gdip.CheckStatus(status);
-        }
-
-        public Metafile(IntPtr henhmetafile, bool deleteEmf)
-        {
-            int status = Gdip.GdipCreateMetafileFromEmf(henhmetafile, deleteEmf, out nativeImage);
-            Gdip.CheckStatus(status);
-        }
-
-        public Metafile(IntPtr referenceHdc, EmfType emfType) :
-            this(referenceHdc, new RectangleF(), MetafileFrameUnit.GdiCompatible, emfType, null)
-        {
-        }
-
-        public Metafile(IntPtr referenceHdc, Rectangle frameRect) :
-            this(referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible, EmfType.EmfPlusDual, null)
-        {
-        }
-
-        public Metafile(IntPtr referenceHdc, RectangleF frameRect) :
-            this(referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible, EmfType.EmfPlusDual, null)
-        {
-        }
-
         public Metafile(IntPtr hmetafile, WmfPlaceableFileHeader wmfHeader)
         {
             int status = Gdip.GdipCreateMetafileFromEmf(hmetafile, false, out nativeImage);
             Gdip.CheckStatus(status);
         }
 
-        public Metafile(Stream stream, IntPtr referenceHdc) :
-            this(stream, referenceHdc, new RectangleF(), MetafileFrameUnit.GdiCompatible, EmfType.EmfPlusDual, null)
-        {
-        }
-
-        public Metafile(string fileName, IntPtr referenceHdc) :
-            this(fileName, referenceHdc, new RectangleF(), MetafileFrameUnit.GdiCompatible, EmfType.EmfPlusDual,
-            null)
-        {
-        }
-
         public Metafile(IntPtr referenceHdc, EmfType emfType, string description) :
             this(referenceHdc, new RectangleF(), MetafileFrameUnit.GdiCompatible, emfType, description)
-        {
-        }
-
-        public Metafile(IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit) :
-            this(referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual, null)
-        {
-        }
-
-        public Metafile(IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit) :
-            this(referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual, null)
-        {
-        }
-
-        public Metafile(IntPtr hmetafile, WmfPlaceableFileHeader wmfHeader, bool deleteWmf)
-        {
-            int status = Gdip.GdipCreateMetafileFromEmf(hmetafile, deleteWmf, out nativeImage);
-            Gdip.CheckStatus(status);
-        }
-
-        public Metafile(Stream stream, IntPtr referenceHdc, EmfType type) :
-            this(stream, referenceHdc, new RectangleF(), MetafileFrameUnit.GdiCompatible, type, null)
-        {
-        }
-
-        public Metafile(Stream stream, IntPtr referenceHdc, Rectangle frameRect) :
-            this(stream, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible, EmfType.EmfPlusDual, null)
-        {
-        }
-
-        public Metafile(Stream stream, IntPtr referenceHdc, RectangleF frameRect) :
-            this(stream, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible, EmfType.EmfPlusDual, null)
-        {
-        }
-
-        public Metafile(string fileName, IntPtr referenceHdc, EmfType type) :
-            this(fileName, referenceHdc, new RectangleF(), MetafileFrameUnit.GdiCompatible, type, null)
-        {
-        }
-
-        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect) :
-            this(fileName, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible, EmfType.EmfPlusDual, null)
-        {
-        }
-
-        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect) :
-            this(fileName, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible, EmfType.EmfPlusDual, null)
-        {
-        }
-
-        public Metafile(IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit, EmfType type) :
-            this(referenceHdc, frameRect, frameUnit, type, null)
-        {
-        }
-
-        public Metafile(IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, EmfType type) :
-            this(referenceHdc, frameRect, frameUnit, type, null)
         {
         }
 
@@ -189,28 +88,8 @@ namespace System.Drawing.Imaging
         {
         }
 
-        public Metafile(Stream stream, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit) :
-            this(stream, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual, null)
-        {
-        }
-
-        public Metafile(Stream stream, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit) :
-            this(stream, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual, null)
-        {
-        }
-
         public Metafile(string fileName, IntPtr referenceHdc, EmfType type, string description) :
             this(fileName, referenceHdc, new RectangleF(), MetafileFrameUnit.GdiCompatible, type, description)
-        {
-        }
-
-        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit) :
-            this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual, null)
-        {
-        }
-
-        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit) :
-            this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual, null)
         {
         }
 
@@ -220,45 +99,6 @@ namespace System.Drawing.Imaging
             int status = Gdip.GdipRecordMetafileI(referenceHdc, type, ref frameRect, frameUnit,
                 desc, out nativeImage);
             Gdip.CheckStatus(status);
-        }
-
-        public Metafile(IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, EmfType type,
-            string description)
-        {
-            int status = Gdip.GdipRecordMetafile(referenceHdc, type, ref frameRect, frameUnit,
-                description, out nativeImage);
-            Gdip.CheckStatus(status);
-        }
-
-        public Metafile(Stream stream, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit,
-            EmfType type) : this(stream, referenceHdc, frameRect, frameUnit, type, null)
-        {
-        }
-
-        public Metafile(Stream stream, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit,
-            EmfType type) : this(stream, referenceHdc, frameRect, frameUnit, type, null)
-        {
-        }
-
-        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit,
-            EmfType type) : this(fileName, referenceHdc, frameRect, frameUnit, type, null)
-        {
-        }
-
-        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit,
-            string description) : this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual, description)
-        {
-        }
-
-        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit,
-            EmfType type) : this(fileName, referenceHdc, frameRect, frameUnit, type, null)
-        {
-        }
-
-        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit,
-            string desc) : this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual,
-            desc)
-        {
         }
 
         public Metafile(Stream stream, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit,
@@ -308,21 +148,6 @@ namespace System.Drawing.Imaging
             int status = Gdip.GdipRecordMetafileFileNameI(fileName, referenceHdc, type, ref frameRect,
                 frameUnit, description, out nativeImage);
             Gdip.CheckStatus(status);
-        }
-
-        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit,
-            EmfType type, string description)
-        {
-            // Called in order to emulate exception behavior from netfx related to invalid file paths.
-            Path.GetFullPath(fileName);
-
-            int status = Gdip.GdipRecordMetafileFileName(fileName, referenceHdc, type, ref frameRect, frameUnit,
-                description, out nativeImage);
-            Gdip.CheckStatus(status);
-        }
-
-        private Metafile(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
         }
 
         // methods
@@ -391,8 +216,8 @@ namespace System.Drawing.Imaging
 
         public static MetafileHeader GetMetafileHeader(string fileName)
         {
-            if (fileName == null)
-                throw new ArgumentNullException(nameof(fileName));
+            // Called in order to emulate exception behavior from netfx related to invalid file paths.
+            Path.GetFullPath(fileName);
 
             IntPtr header = Marshal.AllocHGlobal(Marshal.SizeOf(typeof(MetafileHeader)));
             try
@@ -420,12 +245,6 @@ namespace System.Drawing.Imaging
             {
                 Marshal.FreeHGlobal(header);
             }
-        }
-
-        public void PlayRecord(EmfPlusRecordType recordType, int flags, int dataSize, byte[] data)
-        {
-            int status = Gdip.GdipPlayMetafileRecord(nativeImage, recordType, flags, dataSize, data);
-            Gdip.CheckStatus(status);
         }
     }
 }
