+++ src/System.Drawing.Common/src/System/Drawing/Imaging/Metafile.Windows.cs
@@ -14,50 +14,15 @@ namespace System.Drawing.Imaging
     /// Defines a graphic metafile. A metafile contains records that describe a sequence of graphics operations that
     /// can be recorded and played back.
     /// </summary>
-    [Serializable]
-    [System.Runtime.CompilerServices.TypeForwardedFrom("System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a")]
     public sealed partial class Metafile : Image
     {
-        // GDI+ doesn't handle filenames over MAX_PATH very well
-        private const int MaxPath = 260;
-
         /// <summary>
         /// Initializes a new instance of the <see cref='Metafile'/> class from the specified handle and
         /// <see cref='WmfPlaceableFileHeader'/>.
         /// </summary>
         public Metafile(IntPtr hmetafile, WmfPlaceableFileHeader wmfHeader) :
             this(hmetafile, wmfHeader, false)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified handle and
-        /// <see cref='WmfPlaceableFileHeader'/>.
-        /// </summary>
-        public Metafile(IntPtr hmetafile, WmfPlaceableFileHeader wmfHeader, bool deleteWmf)
         {
-            Gdip.CheckStatus(Gdip.GdipCreateMetafileFromWmf(new HandleRef(null, hmetafile), deleteWmf, wmfHeader, out IntPtr metafile));
-            SetNativeImage(metafile);
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified handle and
-        /// <see cref='WmfPlaceableFileHeader'/>.
-        /// </summary>
-        public Metafile(IntPtr henhmetafile, bool deleteEmf)
-        {
-            Gdip.CheckStatus(Gdip.GdipCreateMetafileFromEmf(new HandleRef(null, henhmetafile), deleteEmf, out IntPtr metafile));
-            SetNativeImage(metafile);
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified filename.
-        /// </summary>
-        public Metafile(string filename)
-        {
-            // Called in order to emulate exception behavior from netfx related to invalid file paths.
-            Path.GetFullPath(filename);
-            Gdip.CheckStatus(Gdip.GdipCreateMetafileFromFile(filename, out IntPtr metafile));
-            SetNativeImage(metafile);
         }
 
         /// <summary>
@@ -66,7 +31,9 @@ namespace System.Drawing.Imaging
         public Metafile(Stream stream)
         {
             if (stream == null)
+            {
                 throw new ArgumentNullException(nameof(stream));
+            }
 
             Gdip.CheckStatus(Gdip.GdipCreateMetafileFromStream(new GPStream(stream), out IntPtr metafile));
             SetNativeImage(metafile);
@@ -75,90 +42,18 @@ namespace System.Drawing.Imaging
         /// <summary>
         /// Initializes a new instance of the <see cref='Metafile'/> class from the specified handle to a device context.
         /// </summary>
-        public Metafile(IntPtr referenceHdc, EmfType emfType) :
-            this(referenceHdc, emfType, null)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified handle to a device context.
-        /// </summary>
         public Metafile(IntPtr referenceHdc, EmfType emfType, string description)
         {
             Gdip.CheckStatus(Gdip.GdipRecordMetafile(
-                new HandleRef(null, referenceHdc),
+                referenceHdc,
                 emfType,
-                NativeMethods.NullHandleRef,
+                IntPtr.Zero,
                 MetafileFrameUnit.GdiCompatible,
                 description,
                 out IntPtr metafile));
 
             SetNativeImage(metafile);
         }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
-        /// by the specified rectangle.
-        /// </summary>
-        public Metafile(IntPtr referenceHdc, RectangleF frameRect) :
-            this(referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
-        /// by the specified rectangle.
-        /// </summary>
-        public Metafile(IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit) :
-            this(referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
-        /// by the specified rectangle.
-        /// </summary>
-        public Metafile(IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, EmfType type) :
-            this(referenceHdc, frameRect, frameUnit, type, null)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
-        /// by the specified rectangle.
-        /// </summary>
-        public Metafile(IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, EmfType type, string description)
-        {
-            Gdip.CheckStatus(Gdip.GdipRecordMetafile(
-                new HandleRef(null, referenceHdc),
-                type,
-                ref frameRect,
-                frameUnit,
-                description,
-                out IntPtr metafile));
-
-            SetNativeImage(metafile);
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
-        /// by the specified rectangle.
-        /// </summary>
-        public Metafile(IntPtr referenceHdc, Rectangle frameRect) :
-            this(referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
-        /// by the specified rectangle.
-        /// </summary>
-        public Metafile(IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit) :
-            this(referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
-        /// by the specified rectangle.
-        /// </summary>
-        public Metafile(IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit, EmfType type) :
-            this(referenceHdc, frameRect, frameUnit, type, null)
-        { }
 
         /// <summary>
         /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
@@ -171,9 +66,9 @@ namespace System.Drawing.Imaging
             if (frameRect.IsEmpty)
             {
                 Gdip.CheckStatus(Gdip.GdipRecordMetafile(
-                    new HandleRef(null, referenceHdc),
+                    referenceHdc,
                     type,
-                    NativeMethods.NullHandleRef,
+                    IntPtr.Zero,
                     MetafileFrameUnit.GdiCompatible,
                     desc,
                     out metafile));
@@ -181,7 +76,7 @@ namespace System.Drawing.Imaging
             else
             {
                 Gdip.CheckStatus(Gdip.GdipRecordMetafileI(
-                    new HandleRef(null, referenceHdc),
+                    referenceHdc,
                     type,
                     ref frameRect,
                     frameUnit,
@@ -195,20 +90,6 @@ namespace System.Drawing.Imaging
         /// <summary>
         /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
         /// </summary>
-        public Metafile(string fileName, IntPtr referenceHdc) :
-            this(fileName, referenceHdc, EmfType.EmfPlusDual, null)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(string fileName, IntPtr referenceHdc, EmfType type) :
-            this(fileName, referenceHdc, type, null)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
         public Metafile(string fileName, IntPtr referenceHdc, EmfType type, string description)
         {
             // Called in order to emulate exception behavior from netfx related to invalid file paths.
@@ -216,94 +97,15 @@ namespace System.Drawing.Imaging
 
             Gdip.CheckStatus(Gdip.GdipRecordMetafileFileName(
                 fileName,
-                new HandleRef(null, referenceHdc),
+                referenceHdc,
                 type,
-                NativeMethods.NullHandleRef,
+                IntPtr.Zero,
                 MetafileFrameUnit.GdiCompatible,
                 description,
                 out IntPtr metafile));
 
             SetNativeImage(metafile);
         }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect) :
-            this(fileName, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit) :
-            this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect,  MetafileFrameUnit frameUnit, EmfType type) :
-            this(fileName, referenceHdc, frameRect, frameUnit, type, null)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, string desc) :
-            this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual, desc)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, EmfType type, string description)
-        {
-            // Called in order to emulate exception behavior from netfx related to invalid file paths.
-            Path.GetFullPath(fileName);
-
-            if (fileName.Length > MaxPath)
-                throw new PathTooLongException();
-
-            Gdip.CheckStatus(Gdip.GdipRecordMetafileFileName(
-                fileName,
-                new HandleRef(null, referenceHdc),
-                type,
-                ref frameRect,
-                frameUnit,
-                description,
-                out IntPtr metafile));
-
-            SetNativeImage(metafile);
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect) :
-            this(fileName, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit) :
-            this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit, EmfType type) :
-            this(fileName, referenceHdc, frameRect, frameUnit, type, null)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit, string description) :
-            this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual, description)
-        { }
 
         /// <summary>
         /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
@@ -319,9 +121,9 @@ namespace System.Drawing.Imaging
             {
                 Gdip.CheckStatus(Gdip.GdipRecordMetafileFileName(
                     fileName,
-                    new HandleRef(null, referenceHdc),
+                    referenceHdc,
                     type,
-                    NativeMethods.NullHandleRef,
+                    IntPtr.Zero,
                     frameUnit,
                     description,
                     out metafile));
@@ -330,7 +132,7 @@ namespace System.Drawing.Imaging
             {
                 Gdip.CheckStatus(Gdip.GdipRecordMetafileFileNameI(
                     fileName,
-                    new HandleRef(null, referenceHdc),
+                    referenceHdc,
                     type,
                     ref frameRect,
                     frameUnit,
@@ -344,27 +146,13 @@ namespace System.Drawing.Imaging
         /// <summary>
         /// Initializes a new instance of the <see cref='Metafile'/> class from the specified data stream.
         /// </summary>
-        public Metafile(Stream stream, IntPtr referenceHdc) :
-            this(stream, referenceHdc, EmfType.EmfPlusDual, null)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified data stream.
-        /// </summary>
-        public Metafile(Stream stream, IntPtr referenceHdc, EmfType type) :
-            this(stream, referenceHdc, type, null)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified data stream.
-        /// </summary>
         public Metafile(Stream stream, IntPtr referenceHdc, EmfType type, string description)
         {
             Gdip.CheckStatus(Gdip.GdipRecordMetafileStream(
                 new GPStream(stream),
-                new HandleRef(null, referenceHdc),
+                referenceHdc,
                 type,
-                NativeMethods.NullHandleRef,
+                IntPtr.Zero,
                 MetafileFrameUnit.GdiCompatible,
                 description,
                 out IntPtr metafile));
@@ -373,34 +161,13 @@ namespace System.Drawing.Imaging
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified data stream.
-        /// </summary>
-        public Metafile(Stream stream, IntPtr referenceHdc, RectangleF frameRect) :
-            this(stream, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(Stream stream, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit) :
-            this(stream, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(Stream stream, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, EmfType type) :
-            this(stream, referenceHdc, frameRect, frameUnit, type, null)
-        { }
-
-        /// <summary>
         /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
         /// </summary>
         public Metafile(Stream stream, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, EmfType type, string description)
         {
             Gdip.CheckStatus(Gdip.GdipRecordMetafileStream(
                 new GPStream(stream),
-                new HandleRef(null, referenceHdc),
+                referenceHdc,
                 type,
                 ref frameRect,
                 frameUnit,
@@ -409,28 +176,7 @@ namespace System.Drawing.Imaging
 
             SetNativeImage(metafile);
         }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified data stream.
-        /// </summary>
-        public Metafile(Stream stream, IntPtr referenceHdc, Rectangle frameRect) :
-            this(stream, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(Stream stream, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit) :
-            this(stream, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
-        { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
-        /// </summary>
-        public Metafile(Stream stream, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit, EmfType type) :
-            this(stream, referenceHdc, frameRect, frameUnit, type, null)
-        { }
-
+        
         /// <summary>
         /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
         /// </summary>
@@ -442,9 +188,9 @@ namespace System.Drawing.Imaging
             {
                 Gdip.CheckStatus(Gdip.GdipRecordMetafileStream(
                     new GPStream(stream),
-                    new HandleRef(null, referenceHdc),
+                    referenceHdc,
                     type,
-                    NativeMethods.NullHandleRef,
+                    IntPtr.Zero,
                     frameUnit,
                     description,
                     out metafile));
@@ -453,7 +199,7 @@ namespace System.Drawing.Imaging
             {
                 Gdip.CheckStatus(Gdip.GdipRecordMetafileStreamI(
                     new GPStream(stream),
-                    new HandleRef(null, referenceHdc),
+                    referenceHdc,
                     type,
                     ref frameRect,
                     frameUnit,
@@ -463,11 +209,6 @@ namespace System.Drawing.Imaging
 
             SetNativeImage(metafile);
         }
-
-        private Metafile(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
-        }
-
 
         /// <summary>
         /// Returns the <see cref='MetafileHeader'/> associated with the specified <see cref='Metafile'/>.
@@ -479,7 +220,7 @@ namespace System.Drawing.Imaging
                 wmf = new MetafileHeaderWmf()
             };
 
-            Gdip.CheckStatus(Gdip.GdipGetMetafileHeaderFromWmf(new HandleRef(null, hmetafile), wmfHeader, header.wmf));
+            Gdip.CheckStatus(Gdip.GdipGetMetafileHeaderFromWmf(hmetafile, wmfHeader, header.wmf));
             return header;
         }
 
@@ -493,7 +234,7 @@ namespace System.Drawing.Imaging
                 emf = new MetafileHeaderEmf()
             };
 
-            Gdip.CheckStatus(Gdip.GdipGetMetafileHeaderFromEmf(new HandleRef(null, henhmetafile), header.emf));
+            Gdip.CheckStatus(Gdip.GdipGetMetafileHeaderFromEmf(henhmetafile, header.emf));
             return header;
         }
 
@@ -634,23 +375,6 @@ namespace System.Drawing.Imaging
         {
             Gdip.CheckStatus(Gdip.GdipGetHemfFromMetafile(new HandleRef(this, nativeImage), out IntPtr hEmf));
             return hEmf;
-        }
-
-        /// <summary>
-        /// Plays an EMF+ file.
-        /// </summary>
-        public void PlayRecord(EmfPlusRecordType recordType, int flags, int dataSize, byte[] data)
-        {
-            // Used in conjunction with Graphics.EnumerateMetafile to play an EMF+
-            // The data must be DWORD aligned if it's an EMF or EMF+.  It must be
-            // WORD aligned if it's a WMF.
-
-            Gdip.CheckStatus(Gdip.GdipPlayMetafileRecord(
-                new HandleRef(this, nativeImage),
-                recordType,
-                flags,
-                dataSize,
-                data));
         }
 
         /// <summary>
