+++ src/System.Drawing.Common/src/System/Drawing/Imaging/Metafile.cs
@@ -1,0 +1,323 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.IO;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using System.Runtime.Serialization;
+using Gdip = System.Drawing.SafeNativeMethods.Gdip;
+
+namespace System.Drawing.Imaging
+{
+    /// <summary>
+    /// Defines a graphic metafile. A metafile contains records that describe a sequence of graphics operations that
+    /// can be recorded and played back.
+    /// </summary>
+    [Serializable]
+    [TypeForwardedFrom("System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a")]
+    public sealed partial class Metafile : Image
+    {
+        // GDI+ doesn't handle filenames over MAX_PATH very well
+        private const int MaxPath = 260;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified handle and
+        /// <see cref='WmfPlaceableFileHeader'/>.
+        /// </summary>
+        public Metafile(IntPtr hmetafile, WmfPlaceableFileHeader wmfHeader, bool deleteWmf)
+        {
+            Gdip.CheckStatus(Gdip.GdipCreateMetafileFromWmf(hmetafile, deleteWmf, wmfHeader, out IntPtr metafile));
+            SetNativeImage(metafile);
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified handle and
+        /// <see cref='WmfPlaceableFileHeader'/>.
+        /// </summary>
+        public Metafile(IntPtr henhmetafile, bool deleteEmf)
+        {
+            Gdip.CheckStatus(Gdip.GdipCreateMetafileFromEmf(henhmetafile, deleteEmf, out IntPtr metafile));
+            SetNativeImage(metafile);
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified filename.
+        /// </summary>
+        public Metafile(string filename)
+        {
+            // Called in order to emulate exception behavior from netfx related to invalid file paths.
+            Path.GetFullPath(filename);
+            Gdip.CheckStatus(Gdip.GdipCreateMetafileFromFile(filename, out IntPtr metafile));
+            SetNativeImage(metafile);
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
+        /// by the specified rectangle.
+        /// </summary>
+        public Metafile(IntPtr referenceHdc, Rectangle frameRect) :
+            this(referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified handle to a device context.
+        /// </summary>
+        public Metafile(IntPtr referenceHdc, EmfType emfType) :
+            this(referenceHdc, emfType, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
+        /// by the specified rectangle.
+        /// </summary>
+        public Metafile(IntPtr referenceHdc, RectangleF frameRect) :
+            this(referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
+        /// by the specified rectangle.
+        /// </summary>
+        public Metafile(IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit) :
+            this(referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
+        /// by the specified rectangle.
+        /// </summary>
+        public Metafile(IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, EmfType type) :
+            this(referenceHdc, frameRect, frameUnit, type, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
+        /// by the specified rectangle.
+        /// </summary>
+        public Metafile(IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, EmfType type, string description)
+        {
+            Gdip.CheckStatus(Gdip.GdipRecordMetafile(
+                referenceHdc,
+                type,
+                ref frameRect,
+                frameUnit,
+                description,
+                out IntPtr metafile));
+
+            SetNativeImage(metafile);
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
+        /// by the specified rectangle.
+        /// </summary>
+        public Metafile(IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit) :
+            this(referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified device context, bounded
+        /// by the specified rectangle.
+        /// </summary>
+        public Metafile(IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit, EmfType type) :
+            this(referenceHdc, frameRect, frameUnit, type, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(string fileName, IntPtr referenceHdc) :
+            this(fileName, referenceHdc, EmfType.EmfPlusDual, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(string fileName, IntPtr referenceHdc, EmfType type) :
+            this(fileName, referenceHdc, type, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect) :
+            this(fileName, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit) :
+            this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect,  MetafileFrameUnit frameUnit, EmfType type) :
+            this(fileName, referenceHdc, frameRect, frameUnit, type, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, string desc) :
+            this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual, desc)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(string fileName, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, EmfType type, string description)
+        {
+            // Called in order to emulate exception behavior from netfx related to invalid file paths.
+            Path.GetFullPath(fileName);
+            if (fileName.Length > MaxPath)
+            {
+                throw new PathTooLongException();
+            }
+
+            Gdip.CheckStatus(Gdip.GdipRecordMetafileFileName(
+                fileName,
+                referenceHdc,
+                type,
+                ref frameRect,
+                frameUnit,
+                description,
+                out IntPtr metafile));
+
+            SetNativeImage(metafile);
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect) :
+            this(fileName, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit) :
+            this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit, EmfType type) :
+            this(fileName, referenceHdc, frameRect, frameUnit, type, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(string fileName, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit, string description) :
+            this(fileName, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual, description)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified data stream.
+        /// </summary>
+        public Metafile(Stream stream, IntPtr referenceHdc) :
+            this(stream, referenceHdc, EmfType.EmfPlusDual, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified data stream.
+        /// </summary>
+        public Metafile(Stream stream, IntPtr referenceHdc, EmfType type) :
+            this(stream, referenceHdc, type, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified data stream.
+        /// </summary>
+        public Metafile(Stream stream, IntPtr referenceHdc, RectangleF frameRect) :
+            this(stream, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(Stream stream, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit) :
+            this(stream, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(Stream stream, IntPtr referenceHdc, RectangleF frameRect, MetafileFrameUnit frameUnit, EmfType type) :
+            this(stream, referenceHdc, frameRect, frameUnit, type, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class from the specified data stream.
+        /// </summary>
+        public Metafile(Stream stream, IntPtr referenceHdc, Rectangle frameRect) :
+            this(stream, referenceHdc, frameRect, MetafileFrameUnit.GdiCompatible)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(Stream stream, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit) :
+            this(stream, referenceHdc, frameRect, frameUnit, EmfType.EmfPlusDual)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref='Metafile'/> class with the specified filename.
+        /// </summary>
+        public Metafile(Stream stream, IntPtr referenceHdc, Rectangle frameRect, MetafileFrameUnit frameUnit, EmfType type) :
+            this(stream, referenceHdc, frameRect, frameUnit, type, null)
+        {
+        }
+
+        private Metafile(SerializationInfo info, StreamingContext context) : base(info, context)
+        {
+        }
+
+        /// <summary>
+        /// Plays an EMF+ file.
+        /// </summary>
+        public void PlayRecord(EmfPlusRecordType recordType, int flags, int dataSize, byte[] data)
+        {
+            // Used in conjunction with Graphics.EnumerateMetafile to play an EMF+
+            // The data must be DWORD aligned if it's an EMF or EMF+.  It must be
+            // WORD aligned if it's a WMF.
+
+            Gdip.CheckStatus(Gdip.GdipPlayMetafileRecord(
+                new HandleRef(this, nativeImage),
+                recordType,
+                flags,
+                dataSize,
+                data));
+        }
+    }
+}
